% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWQR.R
\name{GWQR}
\alias{GWQR}
\title{geographical weighted quantile regression}
\usage{
GWQR(
  dp = F,
  d,
  X,
  y,
  h,
  fig = F,
  tau,
  kernel = F,
  alpha = 20,
  tolar = 0.001,
  process = T,
  adaptive = F,
  iteration = 200
)
}
\arguments{
\item{d}{distance matrix}

\item{X}{input}

\item{y}{output}

\item{h}{bandwidth}

\item{fig}{logic(0 or 1) whether to plot the MSE by bandwidth}

\item{tau}{quantile}

\item{kernel}{kernel function to choose}

\item{alpha}{the step to update bandwidth}

\item{iteration}{}
}
\value{
\code{beta} regression coef
}
\description{
geographical weighted quantile regression
}
