% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWR.R
\name{GWR}
\alias{GWR}
\title{Geographically Weighted Regression}
\usage{
GWR(
  X,
  Y,
  h,
  d,
  maxit = 500,
  kernel = FALSE,
  fig = F,
  dstar = FALSE,
  alpha = 20,
  tolar = 0.001,
  process = T,
  adaptive = F
)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Response variable}

\item{h}{Bandwidth of kernel function}

\item{d}{The distance matrix of sample points}

\item{kernal}{The name of kernel function. only \code{Norm},\code{Binary}, \code{Exp},
\code{bi-square}, \code{rank}, \code{adaptive} can chosen}
}
\value{
\code{beta} regression estimation

\code{Y.fitted}

\code{Weight}

\code{AICc}

\code{MSE}
}
\description{
Geographically Weighted Regression
}
