% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetQuantile.R
\name{SetQuantile}
\alias{SetQuantile}
\title{Set a random quantile from 0 to 1}
\usage{
SetQuantile(tau = TRUE)
}
\arguments{
\item{tau}{a input parameter with always TRUE}
}
\value{
\code{tau} the random quantile at round(x,2)
}
\description{
Set a random quantile from 0 to 1
}
\section{Functions}{
\itemize{
\item \code{SetQuantile()}: is to set a random quantile for a simulation if you have
no idea with setting quantile, return a tau quantile(numeric)

}}
